Adjusted the way data is saved slightly for the case of the Hubbard model.

+++ Keldysh_mfRG/utilities/hdf5_routines.h
@@ -95,7 +95,7 @@ public:
     double * freq_params;
     double * bfreqs_buffer;
     double * ffreqs_buffer;
-    const int self_dim = collapse_all(dimsSE, [](const size_t& a, const size_t& b) -> size_t {return a*b;});    // length of self-energy buffer
+    const size_t self_dim = collapse_all(dimsSE, [](const size_t& a, const size_t& b) -> size_t {return a*b;});    // length of self-energy buffer
 #ifdef KELDYSH_FORMALISM
     const int irred_dim = 16 * n_in;                                  // length of irreducible vertex buffer
 #else
@@ -105,7 +105,7 @@ public:
     h5_comp * irreducible_class;
 
 #if MAX_DIAG_CLASS >= 1
-    const int K1_dim = collapse_all(dimsK1, [](const size_t& a, const size_t& b) -> size_t {return a*b;});                         // length of K1 buffer
+    const size_t K1_dim = collapse_all(dimsK1, [](const size_t& a, const size_t& b) -> size_t {return a*b;});                         // length of K1 buffer
     h5_comp * K1_class_a;
     h5_comp * K1_class_p;
     h5_comp * K1_class_t;
@@ -113,7 +113,7 @@ public:
 #if MAX_DIAG_CLASS >= 2
     double * bfreqs2_buffer;
     double * ffreqs2_buffer;
-    const int K2_dim = collapse_all(dimsK2, [](const size_t& a, const size_t& b) -> size_t {return a*b;});               // length of K2 buffer
+    const size_t K2_dim = collapse_all(dimsK2, [](const size_t& a, const size_t& b) -> size_t {return a*b;});               // length of K2 buffer
     h5_comp * K2_class_a;
     h5_comp * K2_class_p;
     h5_comp * K2_class_t;
@@ -121,7 +121,7 @@ public:
 #if MAX_DIAG_CLASS >= 3
     double * bfreqs3_buffer;
     double * ffreqs3_buffer;
-    const int K3_dim = collapse_all(dimsK3, [](const size_t& a, const size_t& b) -> size_t {return a*b;});    // length of K3 buffer
+    const size_t K3_dim = collapse_all(dimsK3, [](const size_t& a, const size_t& b) -> size_t {return a*b;});    // length of K3 buffer
     h5_comp * K3_class_a;
     h5_comp * K3_class_p;
     h5_comp * K3_class_t;
